pyproject.toml
[project]
name = "mirrorai-core"
version = "1.0.0"
description = "MirrorAI — AI Clone data pipeline, persona builder, and RAG engine"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "chromadb>=0.4.0",
    "langchain>=0.1.0",
    "langchain-community>=0.1.0",
    "sentence-transformers>=2.0.0",
    "underthesea>=6.0.0",
    "scikit-learn>=1.0.0",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
